Remove cflags, ldflags & args before running
In order the solve Issue-55,
When comment line with cflags, ldflags & args is detected it will be cleared before the build & run of the code (only affect the temperate file).

This fix doesn’t affect the interface code and don't remove the line.
The fix just leaves empty line in the temperate file so if the code has other compilation error the line number won't be affected.